requirements.txt
anyio==4.7.0
ariadne==0.24.0
ariadne-django==0.3.0
asgiref==3.8.1
asttokens==2.4.1
black==24.10.0
click==8.1.7
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-stubs-ext==5.1.1
executing==2.1.0
graphql-core==3.2.5
idna==3.10
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.42.0
traitlets==5.14.3
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
wcwidth==0.2.13
